+++ mcp/core/block_store.hpp
@@ -176,8 +176,8 @@ namespace mcp
 		bool transaction_previous_account_state_get(mcp::db::db_transaction & transaction_a, dev::h256 const & link_a, std::vector<h256> & hashs_a);
 		void transaction_previous_account_state_put(mcp::db::db_transaction & transaction_a, dev::h256 const & link_a, std::vector<h256> & hashs_a);
 		
-		bool den_period_mc_get(mcp::db::db_transaction & transaction_a, uint32_t const & hour, mcp::block_hash & hash_a, std::shared_ptr<rocksdb::ManagedSnapshot> snapshot_a = nullptr);
-		void den_period_mc_put(mcp::db::db_transaction & transaction_a, uint32_t const & hour, mcp::block_hash const & hash_a);
+		bool den_period_mc_get(mcp::db::db_transaction & transaction_a, uint64_t const & hour, mcp::block_hash & hash_a, std::shared_ptr<rocksdb::ManagedSnapshot> snapshot_a = nullptr);
+		void den_period_mc_put(mcp::db::db_transaction & transaction_a, uint64_t const & hour, mcp::block_hash const & hash_a);
 
 		void den_ping_get(mcp::db::db_transaction & transaction_a, dev::Address const & addr_a, uint64_t const & hour_a, std::vector<h256> & hashs_a);
 		mcp::db::forward_iterator den_ping_begin(mcp::db::db_transaction & transaction_a, mcp::den_ping_key const & key_a);
